Refactored persistence unit tests to support shared tests for various backends
