+++ common/persistence/persistence-tests/historyPersistenceTest.go
@@ -34,7 +34,7 @@ import (
 )
 
 type (
-	historyPersistenceSuite struct {
+	HistoryPersistenceSuite struct {
 		suite.Suite
 		TestBase
 		// override suite.Suite.Assertions with require.Assertions; this means that s.NotNil(nil) will stop the test,
@@ -43,14 +43,7 @@ type (
 	}
 )
 
-func TestHistoryPersistenceSuite(t *testing.T) {
-	s := new(historyPersistenceSuite)
-	//suite.Run(t, s)
-	s.UseMysql = true
-	suite.Run(t, s)
-}
-
-func (s *historyPersistenceSuite) SetupSuite() {
+func (s *HistoryPersistenceSuite) SetupSuite() {
 	if testing.Verbose() {
 		log.SetOutput(os.Stdout)
 	}
@@ -58,16 +51,16 @@ func (s *historyPersistenceSuite) SetupSuite() {
 	s.SetupWorkflowStore()
 }
 
-func (s *historyPersistenceSuite) SetupTest() {
+func (s *HistoryPersistenceSuite) SetupTest() {
 	// Have to define our overridden assertions in the test setup. If we did it earlier, s.T() will return nil
 	s.Assertions = require.New(s.T())
 }
 
-func (s *historyPersistenceSuite) TearDownSuite() {
+func (s *HistoryPersistenceSuite) TearDownSuite() {
 	s.TearDownWorkflowStore()
 }
 
-func (s *historyPersistenceSuite) TestAppendHistoryEvents() {
+func (s *HistoryPersistenceSuite) TestAppendHistoryEvents() {
 	domainID := "ff03c29f-fcf1-4aea-893d-1a7ec421e3ec"
 	workflowExecution := gen.WorkflowExecution{
 		WorkflowId: common.StringPtr("append-history-events-test"),
@@ -94,7 +87,7 @@ func (s *historyPersistenceSuite) TestAppendHistoryEvents() {
 	s.Nil(err3)
 }
 
-func (s *historyPersistenceSuite) TestGetHistoryEvents() {
+func (s *HistoryPersistenceSuite) TestGetHistoryEvents() {
 	domainID := "0fdc53ef-b890-4870-a944-b9b028ac9742"
 	workflowExecution := gen.WorkflowExecution{
 		WorkflowId: common.StringPtr("get-history-events-test"),
@@ -132,7 +125,7 @@ func newBatchEventForTest(eventIDs []int64, version int64) *testBatchEvent {
 	return &testBatchEvent{batch: batch, events: events}
 }
 
-func (s *historyPersistenceSuite) TestGetHistoryEventsCompatibility() {
+func (s *HistoryPersistenceSuite) TestGetHistoryEventsCompatibility() {
 	domainID := "373de9d6-e41e-42d4-bee9-9e06968e4d0d"
 	workflowExecution := gen.WorkflowExecution{
 		WorkflowId: common.StringPtr("get-history-events-compatibility-test"),
@@ -170,7 +163,7 @@ func (s *historyPersistenceSuite) TestGetHistoryEventsCompatibility() {
 	s.Equal(int64(7), history.Events[1].GetEventId())
 }
 
-func (s *historyPersistenceSuite) TestDeleteHistoryEvents() {
+func (s *HistoryPersistenceSuite) TestDeleteHistoryEvents() {
 	domainID := "373de9d6-e41e-42d4-bee9-9e06968e4d0d"
 	workflowExecution := gen.WorkflowExecution{
 		WorkflowId: common.StringPtr("delete-history-events-test"),
@@ -207,7 +200,7 @@ func (s *historyPersistenceSuite) TestDeleteHistoryEvents() {
 	s.Nil(data1)
 }
 
-func (s *historyPersistenceSuite) TestAppendAndGet() {
+func (s *HistoryPersistenceSuite) TestAppendAndGet() {
 	domainID := uuid.New()
 	workflowExecution := gen.WorkflowExecution{
 		WorkflowId: common.StringPtr("append-and-get-test"),
@@ -236,7 +229,7 @@ func (s *historyPersistenceSuite) TestAppendAndGet() {
 	}
 }
 
-func (s *historyPersistenceSuite) TestOverwriteAndShadowingHistoryEvents() {
+func (s *HistoryPersistenceSuite) TestOverwriteAndShadowingHistoryEvents() {
 	domainID := "003de9c6-e41e-42d4-bee9-9e06968e4d0d"
 	workflowExecution := gen.WorkflowExecution{
 		WorkflowId: common.StringPtr("delete-history-partial-events-test"),
@@ -308,7 +301,7 @@ func (s *historyPersistenceSuite) TestOverwriteAndShadowingHistoryEvents() {
 	}
 }
 
-func (s *historyPersistenceSuite) AppendHistoryEvents(domainID string, workflowExecution gen.WorkflowExecution,
+func (s *HistoryPersistenceSuite) AppendHistoryEvents(domainID string, workflowExecution gen.WorkflowExecution,
 	firstEventID, eventBatchVersion int64, rangeID, txID int64, eventsBatch *persistence.SerializedHistoryEventBatch, overwrite bool) error {
 
 	return s.HistoryMgr.AppendHistoryEvents(&persistence.AppendHistoryEventsRequest{
@@ -323,7 +316,7 @@ func (s *historyPersistenceSuite) AppendHistoryEvents(domainID string, workflowE
 	})
 }
 
-func (s *historyPersistenceSuite) GetWorkflowExecutionHistory(domainID string, workflowExecution gen.WorkflowExecution,
+func (s *HistoryPersistenceSuite) GetWorkflowExecutionHistory(domainID string, workflowExecution gen.WorkflowExecution,
 	firstEventID, nextEventID int64, pageSize int, token []byte) (*gen.History, []byte, error) {
 
 	response, err := s.HistoryMgr.GetWorkflowExecutionHistory(&persistence.GetWorkflowExecutionHistoryRequest{
@@ -342,7 +335,7 @@ func (s *historyPersistenceSuite) GetWorkflowExecutionHistory(domainID string, w
 	return response.History, response.NextPageToken, nil
 }
 
-func (s *historyPersistenceSuite) DeleteWorkflowExecutionHistory(domainID string,
+func (s *HistoryPersistenceSuite) DeleteWorkflowExecutionHistory(domainID string,
 	workflowExecution gen.WorkflowExecution) error {
 
 	return s.HistoryMgr.DeleteWorkflowExecutionHistory(&persistence.DeleteWorkflowExecutionHistoryRequest{
