+++ common/persistence/persistence-tests/metadataPersistenceTest.go
@@ -21,26 +21,27 @@
 package persistencetests
 
 import (
-	"fmt"
 	"os"
+	"testing"
+
+	"fmt"
 	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
-	"testing"
-
-	gen "github.com/uber/cadence/.gen/go/shared"
-	"github.com/uber/cadence/common/cluster"
 
 	"github.com/pborman/uuid"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+
+	gen "github.com/uber/cadence/.gen/go/shared"
+	"github.com/uber/cadence/common/cluster"
 	"github.com/uber/cadence/common/persistence"
 )
 
 type (
-	metadataPersistenceSuiteV2 struct {
+	MetadataPersistenceSuite struct {
 		suite.Suite
 		TestBase
 		// override suite.Suite.Assertions with require.Assertions; this means that s.NotNil(nil) will stop the test,
@@ -49,14 +50,15 @@ type (
 	}
 )
 
-func TestMetadataPersistenceSuiteV2(t *testing.T) {
-	s := new(metadataPersistenceSuiteV2)
+func TestMetadataPersistenceSuite(t *testing.T) {
+	s := new(MetadataPersistenceSuite)
+	//s.UseMysql = false
 	//suite.Run(t, s)
 	s.UseMysql = true
 	suite.Run(t, s)
 }
 
-func (m *metadataPersistenceSuiteV2) SetupSuite() {
+func (m *MetadataPersistenceSuite) SetupSuite() {
 	if testing.Verbose() {
 		log.SetOutput(os.Stdout)
 	}
@@ -64,47 +66,27 @@ func (m *metadataPersistenceSuiteV2) SetupSuite() {
 	m.SetupWorkflowStore()
 }
 
-func (m *metadataPersistenceSuiteV2) SetupTest() {
+func (m *MetadataPersistenceSuite) SetupTest() {
 	// Have to define our overridden assertions in the test setup. If we did it earlier, s.T() will return nil
 	m.Assertions = require.New(m.T())
-
-	// cleanup the domain created
-	var token []byte
-	pageSize := 10
-ListLoop:
-	for {
-		resp, err := m.ListDomains(pageSize, token)
-		m.Nil(err)
-		token = resp.NextPageToken
-		for _, domain := range resp.Domains {
-			m.DeleteDomain(domain.Info.ID, "")
-		}
-		if len(token) == 0 {
-			break ListLoop
-		}
-	}
 }
 
-func (m *metadataPersistenceSuiteV2) TearDownTest() {
-
-}
-
-func (m *metadataPersistenceSuiteV2) TearDownSuite() {
+func (m *MetadataPersistenceSuite) TearDownSuite() {
 	m.TearDownWorkflowStore()
 }
 
-func (m *metadataPersistenceSuiteV2) TestCreateDomain() {
+func (m *MetadataPersistenceSuite) TestCreateDomain() {
 	id := uuid.New()
 	name := "create-domain-test-name"
 	status := persistence.DomainStatusRegistered
 	description := "create-domain-test-description"
 	owner := "create-domain-test-owner"
-	data := map[string]string{"k1": "v1"}
 	retention := int32(10)
 	emitMetric := true
 	isGlobalDomain := false
 	configVersion := int64(0)
 	failoverVersion := int64(0)
+	data := map[string]string{"k1": "v1"}
 
 	resp0, err0 := m.CreateDomain(
 		&persistence.DomainInfo{
@@ -124,6 +106,7 @@ func (m *metadataPersistenceSuiteV2) TestCreateDomain() {
 		configVersion,
 		failoverVersion,
 	)
+
 	m.Nil(err0)
 	m.NotNil(resp0)
 	m.Equal(id, resp0.ID)
@@ -147,7 +130,7 @@ func (m *metadataPersistenceSuiteV2) TestCreateDomain() {
 	m.Equal(configVersion, resp1.ConfigVersion)
 	m.Equal(failoverVersion, resp1.FailoverVersion)
 	m.True(resp1.ReplicationConfig.Clusters[0].ClusterName == cluster.TestCurrentClusterName)
-	m.Equal(persistence.InitialFailoverNotificationVersion, resp1.FailoverNotificationVersion)
+	m.Equal(int64(0), resp1.NotificationVersion)
 
 	resp2, err2 := m.CreateDomain(
 		&persistence.DomainInfo{
@@ -170,10 +153,15 @@ func (m *metadataPersistenceSuiteV2) TestCreateDomain() {
 	m.NotNil(err2)
 	m.IsType(&gen.DomainAlreadyExistsError{}, err2)
 	m.Nil(resp2)
+
+	resp3, err3 := m.GetDomain("", "")
+	m.Nil(resp3)
+	m.IsType(&gen.BadRequestError{}, err3)
 }
 
-func (m *metadataPersistenceSuiteV2) TestGetDomain() {
+func (m *MetadataPersistenceSuite) TestGetDomain() {
 	id := uuid.New()
+	log.Info("uuid in TestGetDomain", id)
 	name := "get-domain-test-name"
 	status := persistence.DomainStatusRegistered
 	description := "get-domain-test-description"
@@ -245,7 +233,7 @@ func (m *metadataPersistenceSuiteV2) TestGetDomain() {
 	m.Equal(isGlobalDomain, resp2.IsGlobalDomain)
 	m.Equal(configVersion, resp2.ConfigVersion)
 	m.Equal(failoverVersion, resp2.FailoverVersion)
-	m.Equal(persistence.InitialFailoverNotificationVersion, resp2.FailoverNotificationVersion)
+	m.Equal(int64(0), resp2.NotificationVersion)
 
 	resp3, err3 := m.GetDomain("", name)
 	m.Nil(err3)
@@ -266,19 +254,15 @@ func (m *metadataPersistenceSuiteV2) TestGetDomain() {
 	m.Equal(isGlobalDomain, resp2.IsGlobalDomain)
 	m.Equal(configVersion, resp2.ConfigVersion)
 	m.Equal(failoverVersion, resp3.FailoverVersion)
-	m.Equal(persistence.InitialFailoverNotificationVersion, resp3.FailoverNotificationVersion)
+	m.Equal(int64(0), resp3.NotificationVersion)
 
 	resp4, err4 := m.GetDomain(id, name)
 	m.NotNil(err4)
 	m.IsType(&gen.BadRequestError{}, err4)
 	m.Nil(resp4)
-
-	resp5, err5 := m.GetDomain("", "")
-	m.Nil(resp5)
-	m.IsType(&gen.BadRequestError{}, err5)
 }
 
-func (m *metadataPersistenceSuiteV2) TestConcurrentCreateDomain() {
+func (m *MetadataPersistenceSuite) TestConcurrentCreateDomain() {
 	id := uuid.New()
 
 	name := "concurrent-create-domain-test-name"
@@ -365,19 +349,20 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentCreateDomain() {
 	m.Equal(true, vi > 0 && vi <= concurrency)
 }
 
-func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
+func (m *MetadataPersistenceSuite) TestConcurrentUpdateDomain() {
 	id := uuid.New()
+	log.Info("uuid in TestConcurrentUpdateDomain", id)
 	name := "concurrent-update-domain-test-name"
 	status := persistence.DomainStatusRegistered
-	description := "update-domain-test-description"
-	owner := "update-domain-test-owner"
-	data := map[string]string{"k1": "v1"}
+	description := "get-domain-test-description"
+	owner := "get-domain-test-owner"
+	data := map[string]string{"k0": "v0"}
 	retention := int32(10)
 	emitMetric := true
 
 	clusterActive := "some random active cluster name"
 	clusterStandby := "some random standby cluster name"
-	configVersion := int64(10)
+	configVersion := int64(11)
 	failoverVersion := int64(59)
 	isGlobalDomain := true
 	clusters := []*persistence.ClusterReplicationConfig{
@@ -388,6 +373,11 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 			ClusterName: clusterStandby,
 		},
 	}
+
+	resp0, err0 := m.GetDomain("", "does-not-exist")
+	m.Nil(resp0)
+	m.NotNil(err0)
+	m.IsType(&gen.EntityNotExistsError{}, err0)
 
 	resp1, err1 := m.CreateDomain(
 		&persistence.DomainInfo{
@@ -411,13 +401,14 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 		failoverVersion,
 	)
 	m.Nil(err1)
+	m.NotNil(resp1)
 	m.Equal(id, resp1.ID)
 
 	resp2, err2 := m.GetDomain(id, "")
 	m.Nil(err2)
-	metadata, err := m.MetadataManagerV2.GetMetadata()
-	m.Nil(err)
-	notificationVersion := metadata.NotificationVersion
+	m.NotNil(resp2)
+	m.Equal(id, resp2.Info.ID)
+	m.Equal(name, resp2.Info.Name)
 
 	concurrency := 16
 	successCount := int32(0)
@@ -445,8 +436,7 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 				},
 				resp2.ConfigVersion,
 				resp2.FailoverVersion,
-				resp2.FailoverNotificationVersion,
-				notificationVersion,
+				resp2.NotificationVersion,
 			)
 			if err3 == nil {
 				atomic.AddInt32(&successCount, 1)
@@ -455,9 +445,11 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 		}(map[string]string{"k0": newValue})
 	}
 	wg.Wait()
+	resp3, err3 := m.GetDomain("", name)
+
+	log.Info("successCount:", successCount)
 	m.Equal(int32(1), successCount)
 
-	resp3, err3 := m.GetDomain("", name)
 	m.Nil(err3)
 	m.NotNil(resp3)
 	m.Equal(id, resp3.Info.ID)
@@ -465,7 +457,6 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 	m.Equal(status, resp3.Info.Status)
 	m.Equal(description, resp3.Info.Description)
 	m.Equal(owner, resp3.Info.OwnerEmail)
-
 	m.Equal(retention, resp3.Config.Retention)
 	m.Equal(emitMetric, resp3.Config.EmitMetric)
 	m.Equal(clusterActive, resp3.ReplicationConfig.ActiveClusterName)
@@ -476,6 +467,7 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 	m.Equal(isGlobalDomain, resp2.IsGlobalDomain)
 	m.Equal(configVersion, resp2.ConfigVersion)
 	m.Equal(failoverVersion, resp3.FailoverVersion)
+	m.Equal(int64(1), resp3.NotificationVersion)
 
 	//check domain data
 	ss := strings.Split(resp3.Info.Data["k0"], "-")
@@ -485,7 +477,7 @@ func (m *metadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 	m.Equal(true, vi > 0 && vi <= concurrency)
 }
 
-func (m *metadataPersistenceSuiteV2) TestUpdateDomain() {
+func (m *MetadataPersistenceSuite) TestUpdateDomain() {
 	id := uuid.New()
 	name := "update-domain-test-name"
 	status := persistence.DomainStatusRegistered
@@ -535,14 +527,9 @@ func (m *metadataPersistenceSuiteV2) TestUpdateDomain() {
 
 	resp2, err2 := m.GetDomain(id, "")
 	m.Nil(err2)
-	metadata, err := m.MetadataManagerV2.GetMetadata()
-	m.Nil(err)
-	notificationVersion := metadata.NotificationVersion
-
 	updatedStatus := persistence.DomainStatusDeprecated
 	updatedDescription := "description-updated"
 	updatedOwner := "owner-updated"
-	//This will overriding the previous key-value pair
 	updatedData := map[string]string{"k1": "v2"}
 	updatedRetention := int32(20)
 	updatedEmitMetric := false
@@ -551,7 +538,6 @@ func (m *metadataPersistenceSuiteV2) TestUpdateDomain() {
 	updateClusterStandby := "other random standby cluster name"
 	updateConfigVersion := int64(12)
 	updateFailoverVersion := int64(28)
-	updateFailoverNotificationVersion := int64(14)
 	updateClusters := []*persistence.ClusterReplicationConfig{
 		{
 			ClusterName: updateClusterActive,
@@ -580,9 +566,9 @@ func (m *metadataPersistenceSuiteV2) TestUpdateDomain() {
 		},
 		updateConfigVersion,
 		updateFailoverVersion,
-		updateFailoverNotificationVersion,
-		notificationVersion,
+		resp2.NotificationVersion,
 	)
+
 	m.Nil(err3)
 
 	resp4, err4 := m.GetDomain("", name)
@@ -603,8 +589,7 @@ func (m *metadataPersistenceSuiteV2) TestUpdateDomain() {
 	}
 	m.Equal(updateConfigVersion, resp4.ConfigVersion)
 	m.Equal(updateFailoverVersion, resp4.FailoverVersion)
-	m.Equal(updateFailoverNotificationVersion, resp4.FailoverNotificationVersion)
-	m.Equal(notificationVersion, resp4.NotificationVersion)
+	m.Equal(resp2.NotificationVersion+1, resp4.NotificationVersion)
 
 	resp5, err5 := m.GetDomain("", name)
 	m.Nil(err5)
@@ -624,11 +609,10 @@ func (m *metadataPersistenceSuiteV2) TestUpdateDomain() {
 	}
 	m.Equal(updateConfigVersion, resp5.ConfigVersion)
 	m.Equal(updateFailoverVersion, resp5.FailoverVersion)
-	m.Equal(updateFailoverNotificationVersion, resp5.FailoverNotificationVersion)
-	m.Equal(notificationVersion, resp5.NotificationVersion)
+	m.Equal(resp2.NotificationVersion+1, resp5.NotificationVersion)
 }
 
-func (m *metadataPersistenceSuiteV2) TestDeleteDomain() {
+func (m *MetadataPersistenceSuite) TestDeleteDomain() {
 	id := uuid.New()
 	name := "delete-domain-test-name"
 	status := persistence.DomainStatusRegistered
@@ -732,113 +716,9 @@ func (m *metadataPersistenceSuiteV2) TestDeleteDomain() {
 	m.Nil(resp9)
 }
 
-func (m *metadataPersistenceSuiteV2) TestListDomains() {
-	clusterActive1 := "some random active cluster name"
-	clusterStandby1 := "some random standby cluster name"
-	clusters1 := []*persistence.ClusterReplicationConfig{
-		{
-			ClusterName: clusterActive1,
-		},
-		{
-			ClusterName: clusterStandby1,
-		},
-	}
-
-	clusterActive2 := "other random active cluster name"
-	clusterStandby2 := "other random standby cluster name"
-	clusters2 := []*persistence.ClusterReplicationConfig{
-		{
-			ClusterName: clusterActive2,
-		},
-		{
-			ClusterName: clusterStandby2,
-		},
-	}
-
-	inputDomains := []*persistence.GetDomainResponse{
-		{
-			Info: &persistence.DomainInfo{
-				ID:          uuid.New(),
-				Name:        "list-domain-test-name-1",
-				Status:      persistence.DomainStatusRegistered,
-				Description: "list-domain-test-description-1",
-				OwnerEmail:  "list-domain-test-owner-1",
-				Data:        map[string]string{"k1": "v1"},
-			},
-			Config: &persistence.DomainConfig{
-				Retention:  109,
-				EmitMetric: true,
-			},
-			ReplicationConfig: &persistence.DomainReplicationConfig{
-				ActiveClusterName: clusterActive1,
-				Clusters:          clusters1,
-			},
-			IsGlobalDomain:  true,
-			ConfigVersion:   133,
-			FailoverVersion: 266,
-		},
-		{
-			Info: &persistence.DomainInfo{
-				ID:          uuid.New(),
-				Name:        "list-domain-test-name-2",
-				Status:      persistence.DomainStatusRegistered,
-				Description: "list-domain-test-description-2",
-				OwnerEmail:  "list-domain-test-owner-2",
-				Data:        map[string]string{"k1": "v2"},
-			},
-			Config: &persistence.DomainConfig{
-				Retention:  326,
-				EmitMetric: false,
-			},
-			ReplicationConfig: &persistence.DomainReplicationConfig{
-				ActiveClusterName: clusterActive2,
-				Clusters:          clusters2,
-			},
-			IsGlobalDomain:  false,
-			ConfigVersion:   400,
-			FailoverVersion: 667,
-		},
-	}
-	for _, domain := range inputDomains {
-		_, err := m.CreateDomain(
-			domain.Info,
-			domain.Config,
-			domain.ReplicationConfig,
-			domain.IsGlobalDomain,
-			domain.ConfigVersion,
-			domain.FailoverVersion,
-		)
-		m.Nil(err)
-	}
-
-	var token []byte
-	pageSize := 1
-	outputDomains := make(map[string]*persistence.GetDomainResponse)
-ListLoop:
-	for {
-		resp, err := m.ListDomains(pageSize, token)
-		m.Nil(err)
-		token = resp.NextPageToken
-		for _, domain := range resp.Domains {
-			outputDomains[domain.Info.ID] = domain
-			// global notification version is already tested, so here we make it 0
-			// so we can test == easily
-			domain.NotificationVersion = 0
-		}
-		if len(token) == 0 {
-			break ListLoop
-		}
-	}
-
-	m.Equal(len(inputDomains), len(outputDomains))
-	for _, domain := range inputDomains {
-		m.Equal(domain, outputDomains[domain.Info.ID])
-	}
-}
-
-func (m *metadataPersistenceSuiteV2) CreateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig,
+func (m *MetadataPersistenceSuite) CreateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig,
 	replicationConfig *persistence.DomainReplicationConfig, isGlobaldomain bool, configVersion int64, failoverVersion int64) (*persistence.CreateDomainResponse, error) {
-	return m.MetadataManagerV2.CreateDomain(&persistence.CreateDomainRequest{
+	return m.MetadataManager.CreateDomain(&persistence.CreateDomainRequest{
 		Info:              info,
 		Config:            config,
 		ReplicationConfig: replicationConfig,
@@ -848,36 +728,28 @@ func (m *metadataPersistenceSuiteV2) CreateDomain(info *persistence.DomainInfo, 
 	})
 }
 
-func (m *metadataPersistenceSuiteV2) GetDomain(id, name string) (*persistence.GetDomainResponse, error) {
-	return m.MetadataManagerV2.GetDomain(&persistence.GetDomainRequest{
+func (m *MetadataPersistenceSuite) GetDomain(id, name string) (*persistence.GetDomainResponse, error) {
+	return m.MetadataManager.GetDomain(&persistence.GetDomainRequest{
 		ID:   id,
 		Name: name,
 	})
 }
 
-func (m *metadataPersistenceSuiteV2) UpdateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig, replicationConfig *persistence.DomainReplicationConfig,
-	configVersion int64, failoverVersion int64, failoverNotificationVersion int64, notificationVersion int64) error {
-	return m.MetadataManagerV2.UpdateDomain(&persistence.UpdateDomainRequest{
-		Info:                        info,
-		Config:                      config,
-		ReplicationConfig:           replicationConfig,
-		ConfigVersion:               configVersion,
-		FailoverVersion:             failoverVersion,
-		FailoverNotificationVersion: failoverNotificationVersion,
-		NotificationVersion:         notificationVersion,
+func (m *MetadataPersistenceSuite) UpdateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig, replicationConfig *persistence.DomainReplicationConfig,
+	configVersion int64, failoverVersion int64, dbVersion int64) error {
+	return m.MetadataManager.UpdateDomain(&persistence.UpdateDomainRequest{
+		Info:                info,
+		Config:              config,
+		ReplicationConfig:   replicationConfig,
+		ConfigVersion:       configVersion,
+		FailoverVersion:     failoverVersion,
+		NotificationVersion: dbVersion,
 	})
 }
 
-func (m *metadataPersistenceSuiteV2) DeleteDomain(id, name string) error {
+func (m *MetadataPersistenceSuite) DeleteDomain(id, name string) error {
 	if len(id) > 0 {
-		return m.MetadataManagerV2.DeleteDomain(&persistence.DeleteDomainRequest{ID: id})
+		return m.MetadataManager.DeleteDomain(&persistence.DeleteDomainRequest{ID: id})
 	}
-	return m.MetadataManagerV2.DeleteDomainByName(&persistence.DeleteDomainByNameRequest{Name: name})
-}
-
-func (m *metadataPersistenceSuiteV2) ListDomains(pageSize int, pageToken []byte) (*persistence.ListDomainsResponse, error) {
-	return m.MetadataManagerV2.ListDomains(&persistence.ListDomainsRequest{
-		PageSize:      pageSize,
-		NextPageToken: pageToken,
-	})
+	return m.MetadataManager.DeleteDomainByName(&persistence.DeleteDomainByNameRequest{Name: name})
 }
