+++ common/persistence/persistence-tests/metadataPersistenceV2Test.go
@@ -21,27 +21,26 @@
 package persistencetests
 
 import (
-	"os"
-	"testing"
-
 	"fmt"
+	"os"
 	"strconv"
 	"strings"
 	"sync"
 	"sync/atomic"
+	"testing"
+
+	gen "github.com/uber/cadence/.gen/go/shared"
+	"github.com/uber/cadence/common/cluster"
 
 	"github.com/pborman/uuid"
 	log "github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
-
-	gen "github.com/uber/cadence/.gen/go/shared"
-	"github.com/uber/cadence/common/cluster"
 	"github.com/uber/cadence/common/persistence"
 )
 
 type (
-	metadataPersistenceSuite struct {
+	MetadataPersistenceSuiteV2 struct {
 		suite.Suite
 		TestBase
 		// override suite.Suite.Assertions with require.Assertions; this means that s.NotNil(nil) will stop the test,
@@ -50,15 +49,14 @@ type (
 	}
 )
 
-func TestMetadataPersistenceSuite(t *testing.T) {
-	s := new(metadataPersistenceSuite)
-	//s.UseMysql = false
+func TestMetadataPersistenceSuiteV2(t *testing.T) {
+	s := new(MetadataPersistenceSuiteV2)
 	//suite.Run(t, s)
 	s.UseMysql = true
 	suite.Run(t, s)
 }
 
-func (m *metadataPersistenceSuite) SetupSuite() {
+func (m *MetadataPersistenceSuiteV2) SetupSuite() {
 	if testing.Verbose() {
 		log.SetOutput(os.Stdout)
 	}
@@ -66,27 +64,47 @@ func (m *metadataPersistenceSuite) SetupSuite() {
 	m.SetupWorkflowStore()
 }
 
-func (m *metadataPersistenceSuite) SetupTest() {
+func (m *MetadataPersistenceSuiteV2) SetupTest() {
 	// Have to define our overridden assertions in the test setup. If we did it earlier, s.T() will return nil
 	m.Assertions = require.New(m.T())
+
+	// cleanup the domain created
+	var token []byte
+	pageSize := 10
+ListLoop:
+	for {
+		resp, err := m.ListDomains(pageSize, token)
+		m.Nil(err)
+		token = resp.NextPageToken
+		for _, domain := range resp.Domains {
+			m.DeleteDomain(domain.Info.ID, "")
+		}
+		if len(token) == 0 {
+			break ListLoop
+		}
+	}
 }
 
-func (m *metadataPersistenceSuite) TearDownSuite() {
+func (m *MetadataPersistenceSuiteV2) TearDownTest() {
+
+}
+
+func (m *MetadataPersistenceSuiteV2) TearDownSuite() {
 	m.TearDownWorkflowStore()
 }
 
-func (m *metadataPersistenceSuite) TestCreateDomain() {
+func (m *MetadataPersistenceSuiteV2) TestCreateDomain() {
 	id := uuid.New()
 	name := "create-domain-test-name"
 	status := persistence.DomainStatusRegistered
 	description := "create-domain-test-description"
 	owner := "create-domain-test-owner"
+	data := map[string]string{"k1": "v1"}
 	retention := int32(10)
 	emitMetric := true
 	isGlobalDomain := false
 	configVersion := int64(0)
 	failoverVersion := int64(0)
-	data := map[string]string{"k1": "v1"}
 
 	resp0, err0 := m.CreateDomain(
 		&persistence.DomainInfo{
@@ -106,7 +124,6 @@ func (m *metadataPersistenceSuite) TestCreateDomain() {
 		configVersion,
 		failoverVersion,
 	)
-
 	m.Nil(err0)
 	m.NotNil(resp0)
 	m.Equal(id, resp0.ID)
@@ -130,7 +147,7 @@ func (m *metadataPersistenceSuite) TestCreateDomain() {
 	m.Equal(configVersion, resp1.ConfigVersion)
 	m.Equal(failoverVersion, resp1.FailoverVersion)
 	m.True(resp1.ReplicationConfig.Clusters[0].ClusterName == cluster.TestCurrentClusterName)
-	m.Equal(int64(0), resp1.NotificationVersion)
+	m.Equal(persistence.InitialFailoverNotificationVersion, resp1.FailoverNotificationVersion)
 
 	resp2, err2 := m.CreateDomain(
 		&persistence.DomainInfo{
@@ -153,15 +170,10 @@ func (m *metadataPersistenceSuite) TestCreateDomain() {
 	m.NotNil(err2)
 	m.IsType(&gen.DomainAlreadyExistsError{}, err2)
 	m.Nil(resp2)
-
-	resp3, err3 := m.GetDomain("", "")
-	m.Nil(resp3)
-	m.IsType(&gen.BadRequestError{}, err3)
 }
 
-func (m *metadataPersistenceSuite) TestGetDomain() {
+func (m *MetadataPersistenceSuiteV2) TestGetDomain() {
 	id := uuid.New()
-	log.Info("uuid in TestGetDomain", id)
 	name := "get-domain-test-name"
 	status := persistence.DomainStatusRegistered
 	description := "get-domain-test-description"
@@ -233,7 +245,7 @@ func (m *metadataPersistenceSuite) TestGetDomain() {
 	m.Equal(isGlobalDomain, resp2.IsGlobalDomain)
 	m.Equal(configVersion, resp2.ConfigVersion)
 	m.Equal(failoverVersion, resp2.FailoverVersion)
-	m.Equal(int64(0), resp2.NotificationVersion)
+	m.Equal(persistence.InitialFailoverNotificationVersion, resp2.FailoverNotificationVersion)
 
 	resp3, err3 := m.GetDomain("", name)
 	m.Nil(err3)
@@ -254,15 +266,19 @@ func (m *metadataPersistenceSuite) TestGetDomain() {
 	m.Equal(isGlobalDomain, resp2.IsGlobalDomain)
 	m.Equal(configVersion, resp2.ConfigVersion)
 	m.Equal(failoverVersion, resp3.FailoverVersion)
-	m.Equal(int64(0), resp3.NotificationVersion)
+	m.Equal(persistence.InitialFailoverNotificationVersion, resp3.FailoverNotificationVersion)
 
 	resp4, err4 := m.GetDomain(id, name)
 	m.NotNil(err4)
 	m.IsType(&gen.BadRequestError{}, err4)
 	m.Nil(resp4)
+
+	resp5, err5 := m.GetDomain("", "")
+	m.Nil(resp5)
+	m.IsType(&gen.BadRequestError{}, err5)
 }
 
-func (m *metadataPersistenceSuite) TestConcurrentCreateDomain() {
+func (m *MetadataPersistenceSuiteV2) TestConcurrentCreateDomain() {
 	id := uuid.New()
 
 	name := "concurrent-create-domain-test-name"
@@ -349,20 +365,19 @@ func (m *metadataPersistenceSuite) TestConcurrentCreateDomain() {
 	m.Equal(true, vi > 0 && vi <= concurrency)
 }
 
-func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
+func (m *MetadataPersistenceSuiteV2) TestConcurrentUpdateDomain() {
 	id := uuid.New()
-	log.Info("uuid in TestConcurrentUpdateDomain", id)
 	name := "concurrent-update-domain-test-name"
 	status := persistence.DomainStatusRegistered
-	description := "get-domain-test-description"
-	owner := "get-domain-test-owner"
-	data := map[string]string{"k0": "v0"}
+	description := "update-domain-test-description"
+	owner := "update-domain-test-owner"
+	data := map[string]string{"k1": "v1"}
 	retention := int32(10)
 	emitMetric := true
 
 	clusterActive := "some random active cluster name"
 	clusterStandby := "some random standby cluster name"
-	configVersion := int64(11)
+	configVersion := int64(10)
 	failoverVersion := int64(59)
 	isGlobalDomain := true
 	clusters := []*persistence.ClusterReplicationConfig{
@@ -373,11 +388,6 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 			ClusterName: clusterStandby,
 		},
 	}
-
-	resp0, err0 := m.GetDomain("", "does-not-exist")
-	m.Nil(resp0)
-	m.NotNil(err0)
-	m.IsType(&gen.EntityNotExistsError{}, err0)
 
 	resp1, err1 := m.CreateDomain(
 		&persistence.DomainInfo{
@@ -401,14 +411,13 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 		failoverVersion,
 	)
 	m.Nil(err1)
-	m.NotNil(resp1)
 	m.Equal(id, resp1.ID)
 
 	resp2, err2 := m.GetDomain(id, "")
 	m.Nil(err2)
-	m.NotNil(resp2)
-	m.Equal(id, resp2.Info.ID)
-	m.Equal(name, resp2.Info.Name)
+	metadata, err := m.MetadataManagerV2.GetMetadata()
+	m.Nil(err)
+	notificationVersion := metadata.NotificationVersion
 
 	concurrency := 16
 	successCount := int32(0)
@@ -436,7 +445,8 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 				},
 				resp2.ConfigVersion,
 				resp2.FailoverVersion,
-				resp2.NotificationVersion,
+				resp2.FailoverNotificationVersion,
+				notificationVersion,
 			)
 			if err3 == nil {
 				atomic.AddInt32(&successCount, 1)
@@ -445,11 +455,9 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 		}(map[string]string{"k0": newValue})
 	}
 	wg.Wait()
-	resp3, err3 := m.GetDomain("", name)
-
-	log.Info("successCount:", successCount)
 	m.Equal(int32(1), successCount)
 
+	resp3, err3 := m.GetDomain("", name)
 	m.Nil(err3)
 	m.NotNil(resp3)
 	m.Equal(id, resp3.Info.ID)
@@ -457,6 +465,7 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 	m.Equal(status, resp3.Info.Status)
 	m.Equal(description, resp3.Info.Description)
 	m.Equal(owner, resp3.Info.OwnerEmail)
+
 	m.Equal(retention, resp3.Config.Retention)
 	m.Equal(emitMetric, resp3.Config.EmitMetric)
 	m.Equal(clusterActive, resp3.ReplicationConfig.ActiveClusterName)
@@ -467,7 +476,6 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 	m.Equal(isGlobalDomain, resp2.IsGlobalDomain)
 	m.Equal(configVersion, resp2.ConfigVersion)
 	m.Equal(failoverVersion, resp3.FailoverVersion)
-	m.Equal(int64(1), resp3.NotificationVersion)
 
 	//check domain data
 	ss := strings.Split(resp3.Info.Data["k0"], "-")
@@ -477,7 +485,7 @@ func (m *metadataPersistenceSuite) TestConcurrentUpdateDomain() {
 	m.Equal(true, vi > 0 && vi <= concurrency)
 }
 
-func (m *metadataPersistenceSuite) TestUpdateDomain() {
+func (m *MetadataPersistenceSuiteV2) TestUpdateDomain() {
 	id := uuid.New()
 	name := "update-domain-test-name"
 	status := persistence.DomainStatusRegistered
@@ -527,9 +535,14 @@ func (m *metadataPersistenceSuite) TestUpdateDomain() {
 
 	resp2, err2 := m.GetDomain(id, "")
 	m.Nil(err2)
+	metadata, err := m.MetadataManagerV2.GetMetadata()
+	m.Nil(err)
+	notificationVersion := metadata.NotificationVersion
+
 	updatedStatus := persistence.DomainStatusDeprecated
 	updatedDescription := "description-updated"
 	updatedOwner := "owner-updated"
+	//This will overriding the previous key-value pair
 	updatedData := map[string]string{"k1": "v2"}
 	updatedRetention := int32(20)
 	updatedEmitMetric := false
@@ -538,6 +551,7 @@ func (m *metadataPersistenceSuite) TestUpdateDomain() {
 	updateClusterStandby := "other random standby cluster name"
 	updateConfigVersion := int64(12)
 	updateFailoverVersion := int64(28)
+	updateFailoverNotificationVersion := int64(14)
 	updateClusters := []*persistence.ClusterReplicationConfig{
 		{
 			ClusterName: updateClusterActive,
@@ -566,9 +580,9 @@ func (m *metadataPersistenceSuite) TestUpdateDomain() {
 		},
 		updateConfigVersion,
 		updateFailoverVersion,
-		resp2.NotificationVersion,
+		updateFailoverNotificationVersion,
+		notificationVersion,
 	)
-
 	m.Nil(err3)
 
 	resp4, err4 := m.GetDomain("", name)
@@ -589,7 +603,8 @@ func (m *metadataPersistenceSuite) TestUpdateDomain() {
 	}
 	m.Equal(updateConfigVersion, resp4.ConfigVersion)
 	m.Equal(updateFailoverVersion, resp4.FailoverVersion)
-	m.Equal(resp2.NotificationVersion+1, resp4.NotificationVersion)
+	m.Equal(updateFailoverNotificationVersion, resp4.FailoverNotificationVersion)
+	m.Equal(notificationVersion, resp4.NotificationVersion)
 
 	resp5, err5 := m.GetDomain("", name)
 	m.Nil(err5)
@@ -609,10 +624,11 @@ func (m *metadataPersistenceSuite) TestUpdateDomain() {
 	}
 	m.Equal(updateConfigVersion, resp5.ConfigVersion)
 	m.Equal(updateFailoverVersion, resp5.FailoverVersion)
-	m.Equal(resp2.NotificationVersion+1, resp5.NotificationVersion)
+	m.Equal(updateFailoverNotificationVersion, resp5.FailoverNotificationVersion)
+	m.Equal(notificationVersion, resp5.NotificationVersion)
 }
 
-func (m *metadataPersistenceSuite) TestDeleteDomain() {
+func (m *MetadataPersistenceSuiteV2) TestDeleteDomain() {
 	id := uuid.New()
 	name := "delete-domain-test-name"
 	status := persistence.DomainStatusRegistered
@@ -716,9 +732,113 @@ func (m *metadataPersistenceSuite) TestDeleteDomain() {
 	m.Nil(resp9)
 }
 
-func (m *metadataPersistenceSuite) CreateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig,
+func (m *MetadataPersistenceSuiteV2) TestListDomains() {
+	clusterActive1 := "some random active cluster name"
+	clusterStandby1 := "some random standby cluster name"
+	clusters1 := []*persistence.ClusterReplicationConfig{
+		{
+			ClusterName: clusterActive1,
+		},
+		{
+			ClusterName: clusterStandby1,
+		},
+	}
+
+	clusterActive2 := "other random active cluster name"
+	clusterStandby2 := "other random standby cluster name"
+	clusters2 := []*persistence.ClusterReplicationConfig{
+		{
+			ClusterName: clusterActive2,
+		},
+		{
+			ClusterName: clusterStandby2,
+		},
+	}
+
+	inputDomains := []*persistence.GetDomainResponse{
+		{
+			Info: &persistence.DomainInfo{
+				ID:          uuid.New(),
+				Name:        "list-domain-test-name-1",
+				Status:      persistence.DomainStatusRegistered,
+				Description: "list-domain-test-description-1",
+				OwnerEmail:  "list-domain-test-owner-1",
+				Data:        map[string]string{"k1": "v1"},
+			},
+			Config: &persistence.DomainConfig{
+				Retention:  109,
+				EmitMetric: true,
+			},
+			ReplicationConfig: &persistence.DomainReplicationConfig{
+				ActiveClusterName: clusterActive1,
+				Clusters:          clusters1,
+			},
+			IsGlobalDomain:  true,
+			ConfigVersion:   133,
+			FailoverVersion: 266,
+		},
+		{
+			Info: &persistence.DomainInfo{
+				ID:          uuid.New(),
+				Name:        "list-domain-test-name-2",
+				Status:      persistence.DomainStatusRegistered,
+				Description: "list-domain-test-description-2",
+				OwnerEmail:  "list-domain-test-owner-2",
+				Data:        map[string]string{"k1": "v2"},
+			},
+			Config: &persistence.DomainConfig{
+				Retention:  326,
+				EmitMetric: false,
+			},
+			ReplicationConfig: &persistence.DomainReplicationConfig{
+				ActiveClusterName: clusterActive2,
+				Clusters:          clusters2,
+			},
+			IsGlobalDomain:  false,
+			ConfigVersion:   400,
+			FailoverVersion: 667,
+		},
+	}
+	for _, domain := range inputDomains {
+		_, err := m.CreateDomain(
+			domain.Info,
+			domain.Config,
+			domain.ReplicationConfig,
+			domain.IsGlobalDomain,
+			domain.ConfigVersion,
+			domain.FailoverVersion,
+		)
+		m.Nil(err)
+	}
+
+	var token []byte
+	pageSize := 1
+	outputDomains := make(map[string]*persistence.GetDomainResponse)
+ListLoop:
+	for {
+		resp, err := m.ListDomains(pageSize, token)
+		m.Nil(err)
+		token = resp.NextPageToken
+		for _, domain := range resp.Domains {
+			outputDomains[domain.Info.ID] = domain
+			// global notification version is already tested, so here we make it 0
+			// so we can test == easily
+			domain.NotificationVersion = 0
+		}
+		if len(token) == 0 {
+			break ListLoop
+		}
+	}
+
+	m.Equal(len(inputDomains), len(outputDomains))
+	for _, domain := range inputDomains {
+		m.Equal(domain, outputDomains[domain.Info.ID])
+	}
+}
+
+func (m *MetadataPersistenceSuiteV2) CreateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig,
 	replicationConfig *persistence.DomainReplicationConfig, isGlobaldomain bool, configVersion int64, failoverVersion int64) (*persistence.CreateDomainResponse, error) {
-	return m.MetadataManager.CreateDomain(&persistence.CreateDomainRequest{
+	return m.MetadataManagerV2.CreateDomain(&persistence.CreateDomainRequest{
 		Info:              info,
 		Config:            config,
 		ReplicationConfig: replicationConfig,
@@ -728,28 +848,36 @@ func (m *metadataPersistenceSuite) CreateDomain(info *persistence.DomainInfo, co
 	})
 }
 
-func (m *metadataPersistenceSuite) GetDomain(id, name string) (*persistence.GetDomainResponse, error) {
-	return m.MetadataManager.GetDomain(&persistence.GetDomainRequest{
+func (m *MetadataPersistenceSuiteV2) GetDomain(id, name string) (*persistence.GetDomainResponse, error) {
+	return m.MetadataManagerV2.GetDomain(&persistence.GetDomainRequest{
 		ID:   id,
 		Name: name,
 	})
 }
 
-func (m *metadataPersistenceSuite) UpdateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig, replicationConfig *persistence.DomainReplicationConfig,
-	configVersion int64, failoverVersion int64, dbVersion int64) error {
-	return m.MetadataManager.UpdateDomain(&persistence.UpdateDomainRequest{
-		Info:                info,
-		Config:              config,
-		ReplicationConfig:   replicationConfig,
-		ConfigVersion:       configVersion,
-		FailoverVersion:     failoverVersion,
-		NotificationVersion: dbVersion,
+func (m *MetadataPersistenceSuiteV2) UpdateDomain(info *persistence.DomainInfo, config *persistence.DomainConfig, replicationConfig *persistence.DomainReplicationConfig,
+	configVersion int64, failoverVersion int64, failoverNotificationVersion int64, notificationVersion int64) error {
+	return m.MetadataManagerV2.UpdateDomain(&persistence.UpdateDomainRequest{
+		Info:                        info,
+		Config:                      config,
+		ReplicationConfig:           replicationConfig,
+		ConfigVersion:               configVersion,
+		FailoverVersion:             failoverVersion,
+		FailoverNotificationVersion: failoverNotificationVersion,
+		NotificationVersion:         notificationVersion,
 	})
 }
 
-func (m *metadataPersistenceSuite) DeleteDomain(id, name string) error {
+func (m *MetadataPersistenceSuiteV2) DeleteDomain(id, name string) error {
 	if len(id) > 0 {
-		return m.MetadataManager.DeleteDomain(&persistence.DeleteDomainRequest{ID: id})
+		return m.MetadataManagerV2.DeleteDomain(&persistence.DeleteDomainRequest{ID: id})
 	}
-	return m.MetadataManager.DeleteDomainByName(&persistence.DeleteDomainByNameRequest{Name: name})
+	return m.MetadataManagerV2.DeleteDomainByName(&persistence.DeleteDomainByNameRequest{Name: name})
+}
+
+func (m *MetadataPersistenceSuiteV2) ListDomains(pageSize int, pageToken []byte) (*persistence.ListDomainsResponse, error) {
+	return m.MetadataManagerV2.ListDomains(&persistence.ListDomainsRequest{
+		PageSize:      pageSize,
+		NextPageToken: pageToken,
+	})
 }
