+++ host/integration_cross_dc_domain_test.go
@@ -22,6 +22,7 @@ package host
 
 import (
 	"flag"
+	"github.com/uber/cadence/common/persistence/persistence-tests"
 	"os"
 	"testing"
 
@@ -91,7 +92,7 @@ func (s *integrationCrossDCSuite) TearDownTest() {
 func (s *integrationCrossDCSuite) setupTest(enableGlobalDomain bool, isMasterCluster bool) {
 	// Have to define our overridden assertions in the test setup. If we did it earlier, s.T() will return nil
 	s.Assertions = require.New(s.T())
-	options := persistence.TestBaseOptions{}
+	options := persistencetests.TestBaseOptions{}
 	options.ClusterHost = "127.0.0.1"
 	options.DropKeySpace = true
 	options.SchemaDir = ".."
