+++ host/integration_test.go
@@ -26,6 +26,7 @@ import (
 	"errors"
 	"flag"
 	"fmt"
+	"github.com/uber/cadence/common/persistence/persistence-tests"
 	"math/rand"
 	"os"
 	"strconv"
@@ -98,7 +99,7 @@ func (s *integrationSuite) TearDownTest() {
 func (s *integrationSuite) setupSuite(enableGlobalDomain bool, isMasterCluster bool) {
 	// Have to define our overridden assertions in the test setup. If we did it earlier, s.T() will return nil
 	s.Assertions = require.New(s.T())
-	options := persistence.TestBaseOptions{}
+	options := persistencetests.TestBaseOptions{}
 	options.ClusterHost = "127.0.0.1"
 	options.DropKeySpace = true
 	options.SchemaDir = ".."
